CSHARP-3523: Split the unified test runner into 2 classes.
