+++ tests/MongoDB.Driver.Tests/Specifications/change-streams/ChangeStreamUnifiedTestRunner.cs
@@ -16,10 +16,10 @@
 using System.Collections.Generic;
 using MongoDB.Bson;
 using MongoDB.Bson.TestHelpers.JsonDrivenTests;
-using MongoDB.Driver.Tests.Specifications.unified_test_format;
+using MongoDB.Driver.Tests.UnifiedTestOperations;
 using Xunit;
 
-namespace MongoDB.Driver.Tests.Specifications.crud.Unified
+namespace MongoDB.Driver.Tests.Specifications.change_streams
 {
     public sealed class ChangeStreamUnifiedTestRunner
     {
@@ -27,7 +27,7 @@ namespace MongoDB.Driver.Tests.Specifications.crud.Unified
         [ClassData(typeof(TestCaseFactory))]
         public void Run(JsonDrivenTestCase testCase)
         {
-            using (var runner = new UnifiedTestFormatTestRunner())
+            using (var runner = new UnifiedTestRunner())
             {
                 runner.Run(testCase);
             }
